Modify: 디테일 하단부 완성
