+++ public/data/detail/vegetable/ecoFriendly/springonion.json
@@ -1,18 +1,39 @@
-{"id": "0101001",
-"category": "0101",
-"nameKor": "친환경 대파 250g",
-"nameEng": "eco friendly springonion",
-"shortDesc": "향이 진하고 단맛도 강한 싱싱한 대파(250g)",
-"imgUrl": "https://img-cf.kurly.com/shop/data/goods/1579511175292l0.jpg",
-"originalPrice": "2190",
-"discountRate": "",
-"salesPrice": "2190",
-"earnPoint": "110",
-"saleUnit": "1봉지",
-"weightOrCapacity": "250g",
-"deliveryCategory": "샛별배송/택배배송",
-"originCountry": "국내",
-"packageType": "냉장/종이포장",
-"packageMessage": "택배배송은 에코포장이 스티로폼으로 대체됩니다.",
-"notification": "-식품 특성상 중량은 3% 내외의 차이가 발생할 수 있습니다.\n-무농약 인증 (제13-6-15 인증기관:오씨케이)\n-해당 상품은 여러 협력업체에서 납품하고 있는 상품으로 수령 시, 상이한 협력업체의 상품을 받아보실 수 있습니다. 품질은 동일 기준으로 입고되고 있사오니 참고 부탁드립니다.\n-2021년 10월 1일 23시 주문 건까지, 제조일자 2021년 9월 28일 상품 출고"
+{
+  "id": "0101001",
+  "category": "0101",
+  "nameKor": "친환경 대파 250g",
+  "nameEng": "eco friendly springonion",
+  "shortDesc": "향이 진하고 단맛도 강한 싱싱한 대파(250g)",
+  "thumbImgUrl": "https://img-cf.kurly.com/shop/data/goods/1579511175292l0.jpg",
+  "originalPrice": 2190,
+  "discountRate": 0,
+  "salesPrice": 2190,
+  "earnPoint": 110,
+  "saleUnit": "1봉지",
+  "weightOrCapacity": "250g",
+  "deliveryCategory": "샛별배송/택배배송",
+  "originCountry": "국내",
+  "packageType": "냉장/종이포장",
+  "packageMessage": "택배배송은 에코포장이 스티로폼으로 대체됩니다.",
+  "notification": "- 식품 특성상 중량은 3% 내외의 차이가 발생할 수 있습니다.\n- 무농약 인증 (제13-6-15 인증기관:오씨케이)\n- 해당 상품은 여러 협력업체에서 납품하고 있는 상품으로 수령 시, 상이한 협력업체의 상품을 받아보실 수 있습니다. 품질은 동일 기준으로 입고되고 있사오니 참고 부탁드립니다.\n- 2021년 10월 1일 23시 주문 건까지, 제조일자 2021년 9월 28일 상품 출고",
+  "detailDesc": {
+    "detailDescImg": "https://img-cf.kurly.com/shop/data/goodsview/20200204/gv00000080249_1.jpg",
+    "detailDescTitle": "싱그러운 대파의 진한 향",
+    "detailDescName": "친환경 대파",
+    "detailDescArticle": "대파는 알싸한 향과 은근한 단맛으로 요리에 감칠맛을 더해주지요. 그때 그때 신선하게 즐기고 싶다면 컬리의 친환경 대파를 만나보세요. 대파 2~3개를 잎과 몸통으로 자른 뒤 깔끔하게 포장했습니다. 무농약으로 재배해 안심할 수 있고, 적당한 크기로 썰려있어 보관하기에도 한결 수월합니다. 초록빛 대파잎과 하얀 잎몸 부분을 섞어 보내드리니, 원하는 대로 손질해 사용하세요."
+  },
+  "additionalInfo": "/image/springonionAddInfo.png",
+  "detailInfo": {
+    "detailInfoImg": "https://img-cf.kurly.com/shop/data/goodsview/20210624/gv20000196772_1.jpg",
+    "productName": "친환경 대파",
+    "certification": "무농약 인증",
+    "packagingUnit": "250g",
+    "matterNote": "해당사항 없음",
+    "producer": "제품 별도 표기 라벨 참조",
+    "howToStore": "냉장 보관",
+    "originCountry": "국내산",
+    "serviceCall": "마켓컬리 고객행복센터 (1644-1107)",
+    "produceDate": "농산물이므로 별도의 유통기한은 없으나 가급적 빨리 드시길 바랍니다."
+  },
+  "customerCenter": "/image.customerCenter.png"
 }
