+++ src/pages/Detail/components/AdditionalInfo.scss
@@ -1,0 +1,3 @@
+.additionalInfoImg {
+  width: 1050px;
+}
