+++ src/pages/Detail/components/BottomLayer.scss
@@ -1,0 +1,207 @@
+@import '../../../styles/variables';
+
+.detailBottomLayer {
+  position: sticky;
+  bottom: 0px;
+  margin-top: 0;
+  margin-bottom: 0;
+  padding-bottom: 20px;
+  border-top: 2px solid $themeColorPrimary;
+  width: 100%;
+  background-color: #fff;
+  text-align: center;
+  z-index: 100;
+
+  .botLayerPopBtn {
+    position: absolute;
+    top: -50px;
+    right: 250px;
+    border: none;
+    border-radius: 2px;
+    width: 170px;
+    height: 50px;
+    background-color: $themeColorPrimary;
+    color: #fff;
+    font-size: 15px;
+    cursor: pointer;
+  }
+
+  .layerDown {
+    display: none;
+  }
+
+  .botLayerSetQuantity {
+    display: inline-block;
+    margin-top: 20px;
+    border-radius: 2px;
+    width: 900px;
+    height: 48px;
+    background-color: #f0f0f0;
+    text-align: left;
+
+    .botLayerProductName {
+      display: inline-block;
+      margin-left: 14px;
+      font-size: 16px;
+    }
+
+    .botLayeQuantitySet {
+      position: relative;
+      display: inline-block;
+      margin-left: 420px;
+
+      .botLayerBtns {
+        display: inline-block;
+        padding-top: 10px;
+        height: 30px;
+
+        .botBtn {
+          overflow: hidden;
+          float: left;
+          width: 28px;
+          height: 28px;
+          border: 0;
+          line-height: 0;
+          text-indent: -14px;
+        }
+
+        .botLayerMinusBtn {
+          border-top: 1px solid #999999;
+          border-left: 1px solid #999999;
+          border-bottom: 1px solid #999999;
+          background: url(https://res.kurly.com/pc/ico/2010/ico_minus_on.svg)
+            no-repeat 50% 50% #fff;
+          border-top-left-radius: 4px;
+          border-bottom-left-radius: 4px;
+          cursor: pointer;
+        }
+
+        .botLayerBtnResult {
+          padding-left: 18px;
+          border-top: 1px solid #999999;
+          border-bottom: 1px solid #999999;
+          background: #fff;
+          text-align: center;
+          font-size: 18px;
+        }
+
+        .botLayerPlusBtn {
+          border-top: 1px solid #999999;
+          border-right: 1px solid #999999;
+          border-bottom: 1px solid #999999;
+          background: url(https://res.kurly.com/pc/ico/2010/ico_plus_on.svg)
+            no-repeat 50% 50% #fff;
+          border-top-right-radius: 4px;
+          border-bottom-right-radius: 4px;
+          cursor: pointer;
+        }
+      }
+
+      .botLayerBtnOriginalPrice {
+        display: inline-block;
+        position: absolute;
+        top: 15px;
+        left: 100px;
+        padding-left: 20px;
+        width: 100%;
+        color: #999999;
+        font-size: 16px;
+        text-decoration: line-through;
+      }
+
+      .botLayerBtnSalesPrice {
+        display: inline-block;
+        position: absolute;
+        top: 15px;
+        left: 200px;
+        padding-left: 10px;
+        width: 100%;
+        font-size: 16px;
+        font-weight: 500;
+      }
+    }
+  }
+
+  .botLayerPriceResult {
+    display: inline-block;
+    margin-top: 30px;
+    width: 900px;
+    text-align: right;
+
+    .botLayerTotalPriceIs {
+      font-size: 14px;
+    }
+
+    .botLayerTotalPrice {
+      padding-left: 5px;
+      font-size: 30px;
+      font-weight: 900;
+    }
+
+    .botLayerTotalPriceWon {
+      font-size: 20px;
+      font-weight: 700;
+    }
+  }
+
+  .botLayerPointResult {
+    display: inline-block;
+    margin-top: 15px;
+    width: 900px;
+    text-align: right;
+
+    .earnPointIcon {
+      display: inline-block;
+      margin-right: 5px;
+      padding-top: 2px;
+      box-sizing: border-box;
+      width: 40px;
+      height: 20px;
+      line-height: 18px;
+      text-align: center;
+      vertical-align: center;
+      border-radius: 8px;
+      background-color: #ffbf00;
+      color: #fff;
+      font-size: 12px;
+    }
+
+    .earnPointDesc {
+      font-size: 16px;
+
+      .earnPointDescStr {
+        font-weight: 900;
+      }
+    }
+  }
+
+  .botLayerCartBtns {
+    display: inline-block;
+    margin-top: 20px;
+    width: 900px;
+    text-align: right;
+
+    .botLayerRestockAlarmBtn {
+      margin-right: 20px;
+      border: 1px solid #dddddd;
+      border-radius: 3px;
+      background-color: #fff;
+      width: 140px;
+      height: 55px;
+      color: #dddddd;
+      font-size: 18px;
+      cursor: pointer;
+    }
+
+    .botLayerCartAddBtn {
+      border: 1px solid #dddddd;
+      border-radius: 3px;
+      background-color: $themeColorPrimary;
+      width: 380px;
+      height: 55px;
+      color: #fff;
+      font-size: 18px;
+      cursor: pointer;
+    }
+  }
+}
