+++ src/pages/Detail/components/CustomerCenter.scss
@@ -1,0 +1,73 @@
+@import '../../../styles/variables';
+
+.customerCenter {
+  text-align: center;
+
+  .customerCenterImg {
+    display: inline-block;
+    margin: 0 auto;
+    width: 100%;
+  }
+
+  .exchangeAndRefundBtnBox {
+    position: relative;
+    width: 1050px;
+    height: 70px;
+    border: 1px solid #dddddd;
+    text-align: left;
+
+    .exchangeAndRefundBtnTitle {
+      display: inline-block;
+      margin-top: 21px;
+      padding: 0 30px;
+      color: $themeColorPrimary;
+      font-size: 22px;
+      font-weight: 700;
+    }
+
+    .exchangeAndRefundBtnDesc {
+      color: #666666;
+      font-size: 18px;
+      font-weight: 100;
+    }
+
+    .exchangeAndRefundBtn {
+      position: absolute;
+      top: 20px;
+      right: 30px;
+      border: none;
+      background-color: #fff;
+      color: $themeColorPrimary;
+      font-size: 22px;
+      cursor: pointer;
+    }
+  }
+
+  .exchangeAndRefundImg {
+    width: 1050px;
+  }
+
+  .deliveryInfo {
+    position: relative;
+    width: 1050px;
+    height: 70px;
+    border: 1px solid #dddddd;
+    text-align: left;
+
+    .deliveryInfoTitle {
+      display: inline-block;
+      margin-top: 21px;
+      padding: 0 30px;
+      color: $themeColorPrimary;
+      font-size: 22px;
+      font-weight: 700;
+    }
+
+    .deliveryInfoDesc {
+      padding-left: 33px;
+      color: #666666;
+      font-size: 18px;
+      font-weight: 100;
+    }
+  }
+}
