+++ src/pages/Detail/components/Desc.scss
@@ -1,0 +1,87 @@
+.detailDescContainer {
+  margin: 0 auto;
+  padding-top: 55px;
+  width: 1050px;
+
+  .detailDescImg {
+    width: 1050px;
+  }
+
+  .detailDescMain {
+    display: block;
+    margin-top: 60px;
+
+    .detailDescTitle {
+      display: block;
+      margin-bottom: 15px;
+      text-align: center;
+      color: #666666;
+      font-size: 32px;
+    }
+
+    .detailDescName {
+      display: block;
+      padding-bottom: 30px;
+      border-bottom: 1px solid #666666;
+      text-align: center;
+      color: #666666;
+      font-size: 50px;
+    }
+
+    .detailDescArticle {
+      display: block;
+      margin-bottom: 100px;
+      padding-top: 30px;
+      color: #666666;
+      line-height: 40px;
+      font-size: 26px;
+      font-weight: 300;
+    }
+  }
+
+  .doolysCheckPoint {
+    .checkPointTitle {
+      display: block;
+      margin-bottom: 50px;
+      color: #666666;
+      text-align: center;
+      font-size: 62px;
+      font-weight: 700;
+    }
+
+    .checkPointTitle::before {
+      display: inline-block;
+      margin: 0 20px 25px 0;
+      width: 190px;
+      height: 1px;
+      background: #666666;
+      content: '';
+    }
+
+    .checkPointTitle::after {
+      display: inline-block;
+      margin: 0 0 25px 20px;
+      width: 190px;
+      height: 1px;
+      background: #666666;
+      content: '';
+    }
+
+    .checkPointImg {
+      padding-bottom: 100px;
+    }
+  }
+
+  .chechCertification {
+    padding-bottom: 200px;
+    text-align: center;
+
+    .certificationImg {
+      width: 950px;
+    }
+  }
+
+  .additionalInfo {
+    width: 1050px;
+  }
+}
