+++ src/pages/Detail/components/DetaiInfo.scss
@@ -1,0 +1,29 @@
+.detailInfo {
+  margin: 0 auto;
+  padding-top: 55px;
+  width: 1050px;
+
+  .infoImg {
+    display: block;
+    margin: 30px auto;
+  }
+
+  .detailInfoList {
+    display: block;
+
+    .infoListTitle {
+      display: inline-block;
+      padding: 15px 0 0 15px;
+      background-color: #f5f5f5;
+      width: 340px;
+      height: 40px;
+      font-size: 13px;
+    }
+
+    .infoListData {
+      display: inline-block;
+      padding-left: 15px;
+      font-size: 13px;
+    }
+  }
+}
