--- src/pages/Detail/components/DetailBottomLayer.js
@@ -1,80 +1,0 @@
-import React, { Component } from 'react';
-import './DetailBottomLayer.scss';
-
-export default class DetailBottomLayer extends Component {
-  constructor() {
-    super();
-    this.state = {
-      isLayerClicked: false,
-    };
-  }
-
-  handleLayer = () => {
-    const { isLayerClicked } = this.state;
-    this.setState({ isLayerClicked: !isLayerClicked });
-  };
-
-  render() {
-    const { isLayerClicked } = this.state;
-    const {
-      nameKor,
-      qauntityMinus,
-      quantity,
-      qauntityPlus,
-      originalPrice,
-      salesPrice,
-      totalPrice,
-      totalEarnPoint,
-    } = this.props;
-
-    return (
-      <div className={'detailBottomLayer'}>
-        <button className='botLayerPopBtn' onClick={this.handleLayer}>
-          상품 선택 {isLayerClicked ? ' ∨' : ' ∧'}
-        </button>
-        <div
-          className={'botLayerContens ' + (isLayerClicked ? '' : 'layerDown')}
-        >
-          <div className='botLayerSetQuantity'>
-            <div className='botLayerProductName'>{nameKor}</div>
-            <div className='botLayeQuantitySet'>
-              <div className='botLayerBtns'>
-                <button
-                  className='botBtn botLayerMinusBtn'
-                  type='button'
-                  onClick={qauntityMinus}
-                />
-                <button className='botBtn botLayerBtnResult'>{quantity}</button>
-                <button
-                  className='botBtn botLayerPlusBtn'
-                  type='button'
-                  onClick={qauntityPlus}
-                />
-              </div>
-              <span className='botLayerBtnOriginalPrice'>
-                {originalPrice} 원
-              </span>
-              <span className='botLayerBtnSalesPrice'>{salesPrice} 원</span>
-            </div>
-          </div>
-          <div className='botLayerPriceResult'>
-            <span className='botLayerTotalPriceIs'>총 상품금액 : </span>
-            <span className='botLayerTotalPrice'>{totalPrice}</span>
-            <span className='botLayerTotalPriceWon'> 원</span>
-          </div>
-          <div className='botLayerPointResult'>
-            <span className='earnPointIcon'>적립</span>
-            <span className='earnPointDesc'>
-              구매 시{' '}
-              <strong className='earnPointDesc'>{totalEarnPoint}원 적립</strong>
-            </span>
-          </div>
-          <div className='botLayerCartBtns'>
-            <button className='botLayerRestockAlarmBtn'>재입고 알림</button>
-            <button className='botLayerCartAddBtn'>장바구니 담기</button>
-          </div>
-        </div>
-      </div>
-    );
-  }
-}
