--- src/pages/Detail/components/DetailInfo.scss
@@ -1,26 +1,0 @@
-.detailInfo {
-  margin: 50px auto;
-  width: 1050px;
-
-  .infoImg {
-    display: block;
-    margin: 30px auto;
-  }
-
-  .detailInfoList {
-    display: block;
-
-    .infoListTitle {
-      display: inline-block;
-      padding: 15px 0 0 15px;
-      background-color: #f5f5f5;
-      width: 340px;
-      height: 40px;
-    }
-
-    .infoListData {
-      display: inline-block;
-      padding-left: 15px;
-    }
-  }
-}
