--- src/pages/Detail/components/DetailScrollBtn.scss
@@ -1,18 +1,0 @@
-@import '../../../styles/variables';
-
-.scrollBtn {
-  display: inline-block;
-  border: 1px solid #eeeeee;
-  outline: 0;
-  width: 262.5px;
-  height: 65px;
-  background-color: #fafafa;
-  font-size: 20px;
-  cursor: pointer;
-}
-
-.clicked {
-  color: $themeColorPrimary;
-  border-bottom: none;
-  background-color: #fff;
-}
