+++ src/utils/addComma.js
@@ -1,0 +1,9 @@
+export default function addComma(amount) {
+  if (typeof amount === 'number') {
+    return amount.toLocaleString();
+  } else if (typeof amount === 'string') {
+    return parseInt(amount).toLocaleString();
+  } else {
+    return '';
+  }
+}
